REFACTOR(diagnostics): restructure module to proper package organization
- Move diagnostics.py from src/ to src/devonboarder/ for proper package structure
- Update imports to use 'from devonboarder import diagnostics'
- Add diagnostics export to devonboarder package __init__.py
- Resolves CI import errors and follows Python packaging best practices
- All 6 diagnostics tests now pass successfully
